backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.8.0
mysql-connector-python==8.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
PyJWT==2.8.0